Reorganize, cleanup, and document global plugin work
This is still very much a checkpiont commit, but fleshed out enough to
give a reasonable sense of the overall architecture.
